Agri_Assistant/Html/requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
pillow
gunicorn
flask-sqlalchemy
flask-bcrypt
PyJWT
google-generativeai
python-dotenv